move auxiliary functions related to aggregation features to own file
